+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixed table cell styling in `markdown` view
 - Added thousandth delimiter for default rendering of a number in `table` view cell
+- Fixed `popup` view to auto hide when trigger elements are removing from a document
 
 ## 1.0.0-beta.62 (14-10-2021)